+++ apps/cold-api/src/platform/modules/resources/surveys/surveys.service.ts
@@ -7,6 +7,7 @@ import { filter, find, map, merge, omit } from 'lodash';
 import { Span } from 'nestjs-ddtrace';
 import { v4 } from 'uuid';
 import { BaseWorker, CacheService, DarklyService, MqttService, PrismaService, SurveyDefinitionsEntity, UpdateSurveyDefinitionsDto, ZodSurveyResponseDto } from '@coldpbc/nest';
+import { SurveyFilterService } from './filter/survey.filter.service';
 
 @Span()
 @Global()
@@ -20,6 +21,7 @@ export class SurveysService extends BaseWorker {
     private prisma: PrismaService,
     private readonly cache: CacheService,
     private readonly mqtt: MqttService,
+    private readonly filterService: SurveyFilterService,
   ) {
     super('SurveysService');
   }
@@ -276,7 +278,7 @@ export class SurveysService extends BaseWorker {
         update: true,
       });
 
-      return def;
+      return this.filterService.filterDependencies(def);
     } catch (e) {
       this.logger.error(e.message, { error: e });
       throw e;